fix(quality): follow-up sonar findings + event/catalog unit tests
- DynamicSolver: extract applySetEvent from handleRoots (S3776)
- Evaluator: DIFFERENTIATE1 constant (S1192); OdeIntegrator: drop
  unnecessary double cast (S1905)
- OdeEventTest covers direction filters, set-action detection, the
  DynamicSystem.Event convenience constructor and the dtable catalog
  entries (the last uncovered new-code lines)

+++ backend/core/src/main/java/com/frees/backend/core/ode/OdeIntegrator.java
@@ -33,7 +33,7 @@ public final class OdeIntegrator {
         double[] times = new double[count];
         for (int i = 0; i < count; i++) {
             times[i] = count == 1 ? tr.endTime
-                    : p.t0() + (tr.endTime - p.t0()) * i / (double) (count - 1);
+                    : p.t0() + (tr.endTime - p.t0()) * i / (count - 1);
         }
         double[][] states = interpolateAt(tr.knotT, tr.knotY, tr.knotF, times);
         return new OdeResult(times, states, tr.recorded, tr.stopped, tr.endTime,
